+++ spec/app/models/short_agent.rb
@@ -1,0 +1,4 @@
+class ShortAgent
+  include Mongoid::Document
+  include Mongoid::Timestamps::Updated::Short
+end