+++ spec/app/models/short_quiz.rb
@@ -1,0 +1,5 @@
+class ShortQuiz
+  include Mongoid::Document
+  include Mongoid::Timestamps::Created::Short
+  field :name, type: String
+end